Add fallback mechanism in git checkout logic
It could be that somewhere down a git tree, there are faulty commits
or missing submodules. Checking out commits which are affected by this
would crash the parsing process. So I implemented a fallback mechanism
that would checkout the last working commit and skips the faulty one.

Not thoroughly tested though!